backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
pandas>=2.0.3
python-dotenv>=0.19.0
SQLAlchemy>=2.0.20
Flask-SQLAlchemy>=3.1.1
marshmallow>=3.20.1
flask-marshmallow>=0.15.0
marshmallow-sqlalchemy>=0.29.0
requests>=2.31.0
google-generativeai>=0.4.0
Werkzeug>=2.0.3 